// src/typings/itemsadder/schemas.d.ts
export type AttributeModifier =
  | {
      value: number;
      operation: "add" | "multiply_base" | "multiply";
      [k: string]: unknown;
    }
  | number;
export type VanillaEnchants =
  | "PROTECTION_ENVIRONMENTAL"
  | "PROTECTION_FIRE"
  | "PROTECTION_FALL"
  | "PROTECTION_EXPLOSIONS"
  | "PROTECTION_PROJECTILE"
  | "OXYGEN"
  | "WATER_WORKER"
  | "THORNS"
  | "DEPTH_STRIDER"
  | "FROST_WALKER"
  | "BINDING_CURSE"
  | "DAMAGE_ALL"
  | "DAMAGE_UNDEAD"
  | "DAMAGE_ARTHROPODS"
  | "KNOCKBACK"
  | "FIRE_ASPECT"
  | "LOOT_BONUS_MOBS"
  | "SWEEPING_EDGE"
  | "DIG_SPEED"
  | "SILK_TOUCH"
  | "DURABILITY"
  | "LOOT_BONUS_BLOCKS"
  | "ARROW_DAMAGE"
  | "ARROW_KNOCKBACK"
  | "ARROW_FIRE"
  | "ARROW_INFINITE"
  | "LUCK"
  | "LURE"
  | "LOYALTY"
  | "IMPALING"
  | "RIPTIDE"
  | "CHANNELING"
  | "MULTISHOT"
  | "QUICK_CHARGE"
  | "PIERCING"
  | "MENDING"
  | "VANISHING_CURSE"
  | "SOUL_SPEED";
export type VanillaItemsAdderCustomBlocks = VanillaItemsAdderCustomBlocks1 & VanillaItemsAdderCustomBlocks2;
// ...
export type VanillaSoundsAndCustom = VanillaSoundsAndCustom1 & VanillaSoundsAndCustom2;
// ...
export type VanillaSoundsAndCustom2 = string;
// ...
export interface Item {
  enabled?: boolean;
  template?: boolean;
  variant_of?: string;
  display_name: string;
  permission_suffix?: string;
  /**
   * Special property to mark item as MMOITEM https://www.spigotmc.org/resources/39267/
   */
  mmoitem?: {
    type?: string;
    id?: string;
    use_mmoitem_displayname?: string;
  };
  /**
   * Resource pack model generator settings.
   * `material` property is required for every item.
   * However custom armors do not require this setting.
   * Custom armors default item material will be leather if not specified.
   */
  resource: {
    material?: VanillaMaterials;
    generate?: boolean;
    model_id?: number;
    model_path?: string;
    [k: string]: unknown;
  };
  lore?: string[];
  /**
   * Slot in which the item has effect
   */
  attribute_modifiers?: {
    mainhand?: AttributeModifiers;
    offhand?: AttributeModifiers;
    head?: AttributeModifiers;
    chest?: AttributeModifiers;
    legs?: AttributeModifiers;
    feet?: AttributeModifiers;
    [k: string]: unknown;
  };
  /**
   * Durability of the item
   */
  durability?: {
    max_custom_durability?: number;
    custom_durability?: number;
    disappear_when_broken?: boolean;
    unbreakable?: boolean;
    usages?: number;
    show_custom_durability_lore?: boolean;
    [k: string]: unknown;
  };
  item_flags?: (
    | "HIDE_ATTRIBUTES"
    | "HIDE_DESTROYS"
    | "HIDE_ENCHANTS"
    | "HIDE_PLACED_ON"
    | "HIDE_POTION_EFFECTS"
    | "HIDE_UNBREAKABLE"
  )[];
  enchants?: (VanillaEnchants | string)[];
  blocked_enchants?: (VanillaEnchants | "ALL")[];
  specific_properties?: SpecificProperties;
  nbt?: string;
  /**
   * Customize some aspects of how the dropped item appears.
   */
  drop?: {
    show_name?: boolean;
    /**
     * You can make an item glowing when dropped on the ground.
     * Very useful for rare items.
     */
    glow?: {
      enabled?: boolean;
      color?:
        | "BLACK"
        | "DARK_BLUE"
        | "DARK_GREEN"
        | "DARK_AQUA"
        | "DARK_RED"
        | "DARK_PURPLE"
        | "GOLD"
        | "GRAY"
        | "DARK_GRAY"
        | "BLUE"
        | "GREEN"
        | "AQUA"
        | "RED"
        | "PURPLE"
        | "YELLOW"
        | "WHITE"
        | "NONE";
      [k: string]: unknown;
    };
    [k: string]: unknown;
  };
  behaviours?: Behaviours;
  events_cooldown?: number;
  /**
   * Special settings for events.
   */
  events_settings?: {
    /**
     * Cooldown settings for this item.
     */
    cooldown?: {
      ticks?: number;
      indicator?: "HIDDEN" | "TITLE" | "BOSSBAR" | "HUD";
      [k: string]: unknown;
    };
    [k: string]: unknown;
  };
  all_actions_permission?: string;
  events?: Events;
  permission?: string;
  [k: string]: unknown;
}
/**
 * These are the vanilla attribute modifiers, you can get more info here https://minecraft.gamepedia.com/Attribute#Attributes_available_on_all_living_entities
 */
export interface AttributeModifiers {
  attackDamage?: AttributeModifier;
  attackSpeed?: AttributeModifier;
  maxHealth?: AttributeModifier;
  movementSpeed?: AttributeModifier;
  armor?: AttributeModifier;
  armorToughness?: AttributeModifier;
  attackKnockback?: AttributeModifier;
  luck?: AttributeModifier;
  [k: string]: unknown;
}
export interface SpecificProperties {
  block?: {
    placed_model: {
      type: "REAL_NOTE" | "REAL_TRANSPARENT" | "REAL_WIRE" | "TILE" | "REAL" | "FIRE";
      break_particles?: "VANILLA" | "BLOCK" | "ITEM";
      break_particles_material?: VanillaItemsAdderCustomBlocks;
      shift_up?: boolean;
      placeable_on_other_real_wire?: boolean;
      placeable_on_water?: boolean;
      placeable_on_lava?: boolean;
      custom_variants?: {
        variant_1?: CustomBlockVariant;
        variant_2?: CustomBlockVariant1;
        variant_3?: CustomBlockVariant2;
        variant_xxx?: CustomBlockVariant3;
        [k: string]: CustomBlockVariant4;
      };
      rotx?: number;
      roty?: number;
      [k: string]: unknown;
    };
    hardness?: number;
    blast_resistance?: number;
    no_explosion?: boolean;
    sound?: {
      break?: {
        name?: VanillaSoundsAndCustom;
        volume?: number;
        pitch?: number;
        [k: string]: unknown;
      };
      place?: {
        name?: VanillaSoundsAndCustom1 & VanillaSoundsAndCustom2;
        volume?: number;
        pitch?: number;
        [k: string]: unknown;
      };
      [k: string]: unknown;
    };
    cancel_drop?: boolean;
    light_level?: number;
    permission_suffix?: {
      break?: string;
      place?: string;
      [k: string]: unknown;
    };
    break_tools_blacklist?: (
      | VanillaItemsAdderCustomItem
      | (
          | "HAND"
          | "hand"
          | "pickaxe"
          | "PICKAXE"
          | "axe"
          | "AXE"
          | "SHOVEL"
          | "shovel"
          | "hoe"
          | "HOE"
          | "sword"
          | "SWORD"
        )
      | {
          [k: string]: string;
        }
    )[];
    break_tools_whitelist?: (
      | VanillaItemsAdderCustomItem3
      | (
          | "HAND"
          | "hand"
          | "pickaxe"
          | "PICKAXE"
          | "axe"
          | "AXE"
          | "SHOVEL"
          | "shovel"
          | "hoe"
          | "HOE"
          | "sword"
          | "SWORD"
        )
      | {
          [k: string]: string;
        }
    )[];
    events_tools_whitelist?: (
      | VanillaItemsAdderCustomItem4
      | (
          | "HAND"
          | "hand"
          | "pickaxe"
          | "PICKAXE"
          | "axe"
          | "AXE"
          | "SHOVEL"
          | "shovel"
          | "hoe"
          | "HOE"
          | "sword"
          | "SWORD"
        )
      | {
          [k: string]: string;
        }
    )[];
    events_tools_blacklist?: (
      | VanillaItemsAdderCustomItem5
      | (
          | "HAND"
          | "hand"
          | "pickaxe"
          | "PICKAXE"
          | "axe"
          | "AXE"
          | "SHOVEL"
          | "shovel"
          | "hoe"
          | "HOE"
          | "sword"
          | "SWORD"
        )
      | {
          [k: string]: string;
        }
    )[];
    [k: string]: unknown;
  };
  bow?: {
    glow_arrows?: boolean;
    fire_arrows?: boolean;
    arrow_damage?: number;
    [k: string]: unknown;
  };
  crossbow?: {
    glow_arrows?: boolean;
    fire_arrows?: boolean;
    arrow_damage?: number;
    multishot?: boolean;
    firework?: boolean;
    [k: string]: unknown;
  };
  seed?: {
    crop_block?:
      | (
          | "WHEAT"
          | "BEETROOT"
          | "CARROTS"
          | "POTATOES"
          | "COCOA"
          | "BAMBOO"
          | "SUGAR_CANE"
          | "SWEET_BERRY_BUSH"
          | "KELP_PLANT"
          | "SEA_PICKLE"
          | "NETHER_WART"
        )
      | string;
    /**
     * List of the possible result drops and their chances
     */
    results?: {
      [k: string]: Result;
    };
    [k: string]: unknown;
  };
  armor?: {
    custom_armor?: string;
    slot?: "HEAD" | "CHEST" | "LEGS" | "FEET" | "head" | "chest" | "legs" | "feet";
    color?:
      | ("ff7e7e" | "75bebe" | "16d4ff" | "530766" | "00ff72" | "010002" | "570e11" | "d36823" | "ff0000")
      | number
      | {
          /**
           * This interface was referenced by `undefined`'s JSON-Schema definition
           * via the `patternProperty` "^(#|)[0-9a-f]{3,6}$".
           */
          [k: string]: string;
        };
    can_wear?: boolean;
    [k: string]: unknown;
  };
  potion?: {
    type?: VanillaPotionTypes;
    color?: Colors;
    /**
     * Attach multiple potion effects to this potion to create a completely custom potion.
     */
    effects?: {
      effect_1?: CustomPotionEffect;
      effect_2?: CustomPotionEffect;
      effect_xxx?: CustomPotionEffect;
      [k: string]: CustomPotionEffect1;
    };
    [k: string]: unknown;
  };
  leather_horse_armor?: {
    color?: Colors;
    [k: string]: unknown;
  };
  book?: {
    author?: string;
    generation?: "ORIGINAL" | "COPY_OF_ORIGINAL" | "COPY_OF_COPY" | "TATTERED";
    pages?: string[];
    [k: string]: unknown;
  };
  [k: string]: unknown;
}
export interface CustomBlockVariant {
  display_name?: string;
  model: string;
  weight?: number;
  x?: number;
  y?: number;
  uvlock?: boolean;
  [k: string]: unknown;
}
export interface CustomBlockVariant1 {
  display_name?: string;
  model: string;
  weight?: number;
  x?: number;
  y?: number;
  uvlock?: boolean;
  [k: string]: unknown;
}
export interface CustomBlockVariant2 {
  display_name?: string;
  model: string;
  weight?: number;
  x?: number;
  y?: number;
  uvlock?: boolean;
  [k: string]: unknown;
}
// ...
export interface Behaviours {
  hat?: boolean;
  keep_on_death?: boolean;
  music_disc?: BehaviourMusicDisc;
  gun?: BehaviourGun;
  furniture_sit?: BehaviourFurnitureSit;
  furniture?: BehaviourFurniture;
  block_trade_machine?: BehaviourTradeMachine;
  furniture_trade_machine?: BehaviourTradeMachine;
  vehicle?: BehaviourVehicle;
  mob?: BehaviourMob;
  liquid_analyzer?: boolean;
  liquid_bucket?: BehaviourLiquidBucket;
  fuel?: BehaviourFuel;
  sapling?: {
    tree_populator?: string;
    [k: string]: unknown;
  };
  [k: string]: unknown;
}
// ...
export interface Events {
  block_break?: Actions;
  attack?: Actions1;
  kill?: Actions2;
  interact?: Interact;
  interact_mainhand?: Interact1;
  interact_offhand?: Interact2;
  drop?: Actions8;
  pickup?: Actions9;
  eat?: Actions10;
  drink?: Actions11;
  bow_shot?: Actions12;
  gun_shot?: Actions13;
  gun_no_ammo?: Actions14;
  gun_reload?: Actions15;
  book_write?: Actions16;
  book_read?: Actions17;
  fishing_start?: Actions18;
  fishing_caught?: Actions19;
  fishing_failed?: Actions20;
  fishing_cancel?: Actions21;
  fishing_bite?: Actions22;
  fishing_in_ground?: Actions23;
  wear?: Actions24;
  unwear?: Actions25;
  held?: Actions26;
  held_offhand?: Actions27;
  unheld?: Actions28;
  unheld_offhand?: Actions29;
  item_throw?: Actions30;
  item_hit_ground?: Actions31;
  item_hit_entity?: Actions32;
  item_break?: Actions33;
  /**
   * Placed Custom block actions
   */
  placed_block?: {
    interact?: Actions34;
    break?: Actions35;
    [k: string]: unknown;
  };
  /**
   * Placed furniture actions
   */
  placed_furniture?: {
    interact?: Actions36;
    break?: Actions37;
    [k: string]: unknown;
  };
  holding_item?: {
    player_damaged_by_entity?: Actions38;
    [k: string]: unknown;
  };
  holding_item_offhand?: {
    player_damaged_by_entity?: Actions39;
    [k: string]: unknown;
  };
  equipped_armor_item?: {
    player_damaged_by_entity?: Actions40;
    [k: string]: unknown;
  };
  bucket_empty?: Actions41;
  bucket_fill?: Actions42;
  /**
   * Events related to vanilla item_frame
   */
  item_frame?: {
    put_item?: Actions43;
    remove_item?: Actions44;
    [k: string]: unknown;
  };
  /**
   * Events related to vanilla armor_stand
   */
  armor_stand?: {
    put_item?: Actions45;
    remove_item?: Actions46;
    [k: string]: unknown;
  };
  [k: string]: unknown;
}